requirements.txt
torch
torchvision
diffusers==0.20.1
transformers
tqdm
imageio
pathlib
argparse
pillow
opencv-python